+++ components/raftstore-v2/src/operation/mod.rs
@@ -24,7 +24,6 @@ pub use ready::{
 };
 
 pub(crate) use self::{
-    bucket::BucketStatsInfo,
     command::SplitInit,
     query::{LocalReader, ReadDelegatePair, SharedReadTablet},
     txn_ext::TxnContext,
