+++ components/raftstore/src/store/worker/mod.rs
@@ -42,7 +42,8 @@ pub use self::{
     },
     region::{Runner as RegionRunner, Task as RegionTask},
     split_check::{
-        Bucket, BucketRange, KeyEntry, Runner as SplitCheckRunner, Task as SplitCheckTask,
+        Bucket, BucketRange, BucketStatsInfo, KeyEntry, Runner as SplitCheckRunner,
+        Task as SplitCheckTask,
     },
     split_config::{
         SplitConfig, SplitConfigManager, BIG_REGION_CPU_OVERLOAD_THRESHOLD_RATIO,
